added DMASws class and various changes to PktSerial and SWS interfaces
DMASWS adds DMA apis to the base SWS class. PktSerial currently uses the DMASWS api's but this should be factored out into two separate classes in the future.

PktSerial has new functions and uses the DMASWS subclass now, which allows the specification of a DMA callback, removing the specific behaviours from the main SWS driver.

A new Mode was added to the SWS class - Disconnected, which should be used to unconfigure uart after a transaction.