Each packet can now have different communication rates.
More capable devices will always run at 1mbaud, that is until a lower class device joins the bus that can't communicate as fast. At this point, if a virtual driver is created, the high class device will need to remember the communication baud rate of the lower class device so that it can communicate data if required.

This commit tracks the communication rate of devices on the bus, and replaces JDProtocol::send with an instance based JDDriver->send.