CRC is now computed in the control service.
Previously the crc was calculated at the physical layer. This meant that packets not intended for the device would be ignored which is not ideal if a bridge service is operating as a bridge should see all packets.

It makes much more sense to receive all packets and filter them at the control layer anyway.